+++ starters/nextjs-starter-approuter-ts/app/api/pantheoncloud/[...command]/api-options.ts
@@ -1,5 +1,9 @@
 import { PantheonAPIOptions } from "@pantheon-systems/pcc-react-sdk/server";
 import { serverSmartComponentMap } from "../../../../components/smart-components/server-components";
+import {
+  getAuthorById,
+  listAuthors,
+} from "../../../../lib/pcc-metadata-groups";
 
 export const pantheonAPIOptions: PantheonAPIOptions = {
   resolvePath: (article) => `/articles/${article.slug || article.id}`,
@@ -7,4 +11,16 @@ export const pantheonAPIOptions: PantheonAPIOptions = {
   smartComponentMap: serverSmartComponentMap,
   componentPreviewPath: (componentName) =>
     `/component-preview/${componentName}`,
+  metadataGroups: [
+    {
+      label: "Author",
+      groupIdentifier: "AUTHOR",
+      schema: {
+        name: "string",
+        image: "file",
+      },
+      get: getAuthorById,
+      list: listAuthors,
+    },
+  ],
 };